Selecting files from the 'general document repository' and printing algorithm.
This algorithm was written in a way that does not interfere with printer permissions. Since the order of the printouts is important for what I need, a three-second timeout has been added after each printing operation, taking into account the time it takes for the printer to accept documents of different sizes into its memory.